Navigation test
Navigation item "New here?" is different to the page title it links to.
It's also less descriptive of the content on the other side of the link
than "Who We Are", the title of the destination.

Testing the navigational title of "Who We Are", to see if changes usage
habits.

Changes to be committed:
modified:   _includes/more.html
modified:   _includes/navigation.html