Update contact form variables and names
The contact forms were still using 'phone' as a variable for email
fields, from previous versions where the phone number was the main
form capture. Here we're fixing that by calling them 'email' again, in
all the areas it appears:

* Involved item forms
* Map connect forms ("simple" contact forms)
* Contact page form

The javascript that sends the variables through to the mailer app still
sends it through as 'detail' with other things, so the mailer app
doesn't need updating for this update to take effect. The javascript was
updated to take care of the changes all on this side.