Add contact form user variable to Post function
The javascript that sends data to the mailer app didn't send the sender
email address in the raw, as it wasn't previously needed. In order to
send a thank you email, the that user email was captured with the 'user'
variable, and sent with the rest.

The mailer app can now create thank you emails with this information.