Update TIO styles
The stack of links for bible quotes was getting quite long on some
pages, and it was causing the printable versions to fall onto two pages
in some occasions.

This update takes those bible links and stacks them inline, with space
for easy reading.

This also keeps those printable versions on one page.